Move pin-related types into their own module
Pins don't can't really be assigned to a single peripheral, like many
other concepts in this HAL. GPIO, SWM, and IOCON all do things related
to them.

Some time ago, I decided to move them to the `swm` module, for lack of a
better idea. I believe this wasn't a good decision. They belong in their
own module. In my opinion, this is the least confusing.

There might be more things we can do to improve things, but I think this
is already a good first step.
